// packages/types/types/enums.ts
export enum SIDE {
    SELL = "SELL",
    BUY = "BUY",
}

export enum ORDER_TYPE {
    LIMIT = "LIMIT",
    MARKET = "MARKET",
}

export enum MarketStatus {
    OPEN = "OPEN",
    PAUSED = "PAUSED",
    CLOSED = "CLOSED",
}

export enum ORDER_STATUS {
    NEW = "NEW",
    FILLED = "FILLED",
    CANCELLED = "CANCELLED",
    OPEN = "OPEN",
}